+++ p_mobj.c
@@ -512,6 +512,13 @@ floater:
 	  return;
 	}
     }
+  else if (mo->flags2 & MF2_LOGRAV)    // mbf21: low gravity
+    {
+      if (!mo->momz)
+        mo->momz = -(gravity >> 3) * 2;
+      else
+        mo->momz -= gravity >> 3;
+    }
   else // still above the floor
     if (!(mo->flags & MF_NOGRAVITY))
       {
@@ -699,6 +706,7 @@ mobj_t *P_SpawnMobj(fixed_t x, fixed_t y, fixed_t z, mobjtype_t type)
   mobj->radius = info->radius;
   mobj->height = info->height;                                      // phares
   mobj->flags  = info->flags;
+  mobj->flags2 = info->flags2;   // mbf21
   mobj->skin = NULL;
 
   // killough 8/23/98: no friends, bouncers, or touchy things in old demos